+++ app/src/main/AndroidManifest.xml
@@ -12,8 +12,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.BiciMadStations">
-        <activity android:name=".ui.DetailActivity"></activity>
-        <activity android:name=".ui.MainActivity">
+        <activity android:name=".ui.detail.DetailActivity"></activity>
+        <activity android:name=".ui.main.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
